+++ Simulation.Scenario/Mesh.cs
@@ -5,7 +5,7 @@ namespace Simulation
     public class Mesh
     {
         public required Vertex[] Vertices { get; init; }
-        public required short[] Indices { get; init; }
+        public required uint[] Indices { get; init; }
 
         public static readonly Vector3 Facing = new Vector3(0, 0, 1);
     }
